backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
whisper
python-dotenv
